requirements.txt
streamlit>=1.35.0
plotly>=5.20.0
pandas>=2.0.0
python-dotenv>=1.0.0

# S3 連接（接通真實資料時需要）
boto3>=1.34.0

# Google Sheets（如果 Inventory 在 GSheet）
gspread>=6.0.0
google-auth>=2.28.0

# 即時報價 API（選配，接通即時價格時需要）
# tiingo       # pip install tiingo
# requests     # 已內建於 Python 標準庫
